requirements.txt
autopep8==1.5.7
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
tifffile==2021.4.8
toml==0.10.2
